requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
cairocffi>=1.6.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi>=1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cssselect2==0.7.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
Django==5.1.3
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.6
jiter==0.8.2
joblib==1.4.2
langdetect==1.0.9
llama-cloud==0.1.12
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.7
llama-index-llms-groq==0.3.1
llama-index-llms-openai==0.3.20
llama-index-llms-openai-like==0.3.3
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
marshmallow==3.26.1
mocodo==4.2.4
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy>=1.26.2
openai==1.63.2
packaging==24.2
pandas>=2.2.3
pillow>=10.2.0
platformdirs==4.3.6
propcache==0.2.1
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.2
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.21.0
tqdm==4.67.1
transformers>=4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.28.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.18.3
